bot/requirements.txt
python-telegram-bot>=20.0
anthropic>=0.34.0
python-dotenv>=1.0.0
google-cloud-aiplatform>=1.38.0
google-genai>=0.3.0
google-auth>=2.23.0
requests>=2.31.0
beautifulsoup4>=4.12.0
aiohttp>=3.9.0
google-api-python-client>=2.100.0
notion-client>=2.2.1

